sepearted smbus class
